extract utils function: `extractHoleFromPolygonGeometry`
