+++ interactions/pickholesinteraction.js
@@ -1,32 +1,10 @@
-const {
-  dissolve,
-} = g3wsdk.core.geoutils;
-
-const { Geometry } = g3wsdk.core.geometry;
-
 /**
- * Util function to extract hole feature from polygon Geometry
- * @param geometry
- * @param id
- * @param index
- * @returns Array of hole features
+ * @since g3w-client-plugin-editing@v3.7.0
  */
-function extractHoleFromPolygonGeometry({geometry, id, index}={}) {
-  const holesFeatures = [];
-  const linearRingCount = geometry.getLinearRingCount();
-  if ( linearRingCount > 1) {
-    for (let i = 1; i < linearRingCount; i++) {
-      holesFeatures.push(
-        new ol.Feature({
-          geometry: new ol.geom.Polygon([geometry.getLinearRing(i).getCoordinates()]),
-          __id:     id,
-          __index:  index
-        })
-      )
-    }
-  }
-  return holesFeatures;
-}
+import { extractHoleFromPolygonGeometry } from '../utils/extractHoleFromPolygonGeometry';
+
+const { dissolve } = g3wsdk.core.geoutils;
+const { Geometry } = g3wsdk.core.geometry;
 
 const PickHolesEventType = {
   PICKED: 'picked'
